refactor(extraction): split grounding rules
